feat(firebase): fix 2 versions of proxy and extract to own file
